+++ cloud-core/src/main/java/cloud/commandframework/arguments/preprocessor/RegexPreprocessor.java
@@ -112,6 +112,7 @@ public final class RegexPreprocessor<C> implements BiFunction<@NonNull CommandCo
     /**
      * Exception thrown when input fails regex matching in {@link RegexPreprocessor}
      */
+    @SuppressWarnings("serial")
     public static final class RegexValidationException extends IllegalArgumentException {
 
         private static final long serialVersionUID = 747826566058072233L;